renamed reducer file to index
